refactor scaffold bootstrap package helpers
