+++ .sampo/changesets/issue-357-scaffold-bootstrap-package-helpers.md
@@ -1,0 +1,5 @@
+---
+npm/@wp-typia/project-tools: patch
+---
+
+Split scaffold bootstrap and package-manager normalization helpers into focused runtime modules without changing generated project behavior.
