+++ packages/wp-typia-project-tools/src/runtime/scaffold-package-manager-files.ts
@@ -1,0 +1,124 @@
+import fs from "node:fs";
+import { promises as fsp } from "node:fs";
+import { execSync } from "node:child_process";
+import path from "node:path";
+
+import {
+	formatInstallCommand,
+	getPackageManager,
+	transformPackageManagerText,
+} from "./package-managers.js";
+import type { PackageManagerId } from "./package-managers.js";
+
+const LOCKFILES: Record<PackageManagerId, string[]> = {
+	bun: ["bun.lock", "bun.lockb"],
+	npm: ["package-lock.json"],
+	pnpm: ["pnpm-lock.yaml"],
+	yarn: ["yarn.lock"],
+};
+
+interface GeneratedPackageJson {
+	packageManager?: string;
+	scripts?: Record<string, string>;
+}
+
+/**
+ * Normalizes scaffolded package-manager specific files for the selected toolchain.
+ *
+ * @param targetDir Absolute scaffold target directory.
+ * @param packageManagerId Selected package manager id.
+ * @returns A promise that resolves after any package-manager sidecar files are updated.
+ */
+export async function normalizePackageManagerFiles(
+	targetDir: string,
+	packageManagerId: PackageManagerId,
+): Promise<void> {
+	const yarnRcPath = path.join(targetDir, ".yarnrc.yml");
+
+	if (packageManagerId === "yarn") {
+		await fsp.writeFile(yarnRcPath, "nodeLinker: node-modules\n", "utf8");
+		return;
+	}
+
+	if (fs.existsSync(yarnRcPath)) {
+		await fsp.rm(yarnRcPath, { force: true });
+	}
+}
+
+/**
+ * Rewrites the generated package.json for the selected package manager.
+ *
+ * @param targetDir Absolute scaffold target directory.
+ * @param packageManagerId Selected package manager id.
+ * @returns A promise that resolves after the package.json file is normalized.
+ */
+export async function normalizePackageJson(
+	targetDir: string,
+	packageManagerId: PackageManagerId,
+): Promise<void> {
+	const packageJsonPath = path.join(targetDir, "package.json");
+	if (!fs.existsSync(packageJsonPath)) {
+		return;
+	}
+
+	const packageManager = getPackageManager(packageManagerId);
+	const packageJson = JSON.parse(await fsp.readFile(packageJsonPath, "utf8")) as GeneratedPackageJson;
+	packageJson.packageManager = packageManager.packageManagerField;
+
+	if (packageJson.scripts) {
+		for (const [key, value] of Object.entries(packageJson.scripts)) {
+			if (typeof value === "string") {
+				packageJson.scripts[key] = transformPackageManagerText(value, packageManagerId);
+			}
+		}
+	}
+
+	await fsp.writeFile(packageJsonPath, `${JSON.stringify(packageJson, null, "\t")}\n`, "utf8");
+}
+
+/**
+ * Removes lockfiles that do not match the selected package manager.
+ *
+ * @param targetDir Absolute scaffold target directory.
+ * @param packageManagerId Selected package manager id.
+ * @returns A promise that resolves after stale lockfiles are removed.
+ */
+export async function removeUnexpectedLockfiles(
+	targetDir: string,
+	packageManagerId: PackageManagerId,
+): Promise<void> {
+	const keep = new Set(LOCKFILES[packageManagerId] ?? []);
+	const allLockfiles = Object.values(LOCKFILES).flat();
+
+	await Promise.all(
+		allLockfiles.map(async (filename) => {
+			if (keep.has(filename)) {
+				return;
+			}
+
+			const filePath = path.join(targetDir, filename);
+			if (fs.existsSync(filePath)) {
+				await fsp.rm(filePath, { force: true });
+			}
+		}),
+	);
+}
+
+/**
+ * Installs scaffolded project dependencies with the selected package manager.
+ *
+ * @param options Absolute target directory and selected package manager id.
+ * @returns A promise that resolves after the install command completes.
+ */
+export async function defaultInstallDependencies({
+	projectDir,
+	packageManager,
+}: {
+	projectDir: string;
+	packageManager: PackageManagerId;
+}): Promise<void> {
+	execSync(formatInstallCommand(packageManager), {
+		cwd: projectDir,
+		stdio: "inherit",
+	});
+}
