+++ packages/wp-typia-project-tools/src/runtime/scaffold.ts
@@ -5,10 +5,7 @@ import { execSync } from "node:child_process";
 
 import {
 	PACKAGE_MANAGER_IDS,
-	formatPackageExecCommand,
-	formatInstallCommand,
 	getPackageManager,
-	transformPackageManagerText,
 } from "./package-managers.js";
 import type { PackageManagerId } from "./package-managers.js";
 import {
@@ -35,12 +32,18 @@ import {
 import { applyMigrationUiCapability } from "./migration-ui-capability.js";
 import { getPackageVersions } from "./package-versions.js";
 import {
-	ensureMigrationDirectories,
-	writeInitialMigrationScaffold,
-	writeMigrationConfig,
-} from "./migration-project.js";
-import { syncPersistenceRestArtifacts } from "./persistence-rest-artifacts.js";
-import { getStarterManifestFiles, stringifyStarterManifest } from "./starter-manifests.js";
+	applyWorkspaceMigrationCapability,
+	ensureScaffoldDirectory,
+	isOfficialWorkspaceProject,
+	seedBuiltInPersistenceArtifacts,
+	writeStarterManifestFiles,
+} from "./scaffold-bootstrap.js";
+import {
+	defaultInstallDependencies,
+	normalizePackageJson,
+	normalizePackageManagerFiles,
+	removeUnexpectedLockfiles,
+} from "./scaffold-package-manager-files.js";
 import {
 	toPascalCase,
 	toSnakeCase,
@@ -56,12 +59,10 @@ import {
 import { copyInterpolatedDirectory } from "./template-render.js";
 import {
 	OFFICIAL_WORKSPACE_TEMPLATE_PACKAGE,
-	PROJECT_TOOLS_PACKAGE_ROOT,
 	TEMPLATE_IDS,
 	getTemplateById,
 	isBuiltInTemplateId,
 } from "./template-registry.js";
-import type { BuiltInTemplateId } from "./template-registry.js";
 import { resolveTemplateSource } from "./template-source.js";
 import {
 	BlockGeneratorService,
@@ -70,13 +71,6 @@ import {
 } from "./block-generator-service.js";
 
 const WORKSPACE_TEMPLATE_ALIAS = "workspace";
-const EPHEMERAL_NODE_MODULES_LINK_TYPE = process.platform === "win32" ? "junction" : "dir";
-const LOCKFILES: Record<PackageManagerId, string[]> = {
-	bun: ["bun.lock", "bun.lockb"],
-	npm: ["package-lock.json"],
-	pnpm: ["pnpm-lock.yaml"],
-	yarn: ["yarn.lock"],
-};
 
 /**
  * User-facing scaffold answers before template rendering.
@@ -228,15 +222,6 @@ export interface ScaffoldProjectResult {
 	templateId: string;
 	variables: ScaffoldTemplateVariables;
 	warnings: string[];
-}
-
-interface GeneratedPackageJson {
-	dependencies?: Record<string, string>;
-	scripts?: Record<string, string>;
-	wpTypia?: {
-		projectType?: string;
-		templatePackage?: string;
-	};
 }
 export { buildBlockCssClassName } from "./scaffold-identifiers.js";
 
@@ -519,268 +504,6 @@ export function getTemplateVariables(
 	};
 }
 
-async function ensureDirectory(targetDir: string, allowExisting = false): Promise<void> {
-	if (!fs.existsSync(targetDir)) {
-		await fsp.mkdir(targetDir, { recursive: true });
-		return;
-	}
-
-	if (allowExisting) {
-		return;
-	}
-
-	const entries = await fsp.readdir(targetDir);
-	if (entries.length > 0) {
-		throw new Error(`Target directory is not empty: ${targetDir}`);
-	}
-}
-
-async function writeStarterManifestFiles(
-	targetDir: string,
-	templateId: string,
-	variables: ScaffoldTemplateVariables,
-): Promise<void> {
-	const manifests = getStarterManifestFiles(templateId, variables);
-
-	for (const { document, relativePath } of manifests) {
-		const destinationPath = path.join(targetDir, relativePath);
-		await fsp.mkdir(path.dirname(destinationPath), { recursive: true });
-		await fsp.writeFile(destinationPath, stringifyStarterManifest(document), "utf8");
-	}
-}
-
-/**
- * Seed REST-derived persistence artifacts into a newly scaffolded built-in
- * project before the first manual `sync-rest` run.
- *
- * @param targetDir Absolute scaffold target directory.
- * @param templateId Built-in template id being scaffolded.
- * @param variables Resolved scaffold template variables for the project.
- * @returns A promise that resolves after any required persistence artifacts are generated.
- */
-async function seedBuiltInPersistenceArtifacts(
-	targetDir: string,
-	templateId: BuiltInTemplateId,
-	variables: ScaffoldTemplateVariables,
-): Promise<void> {
-	const needsPersistenceArtifacts =
-		templateId === "persistence" ||
-		(templateId === "compound" && variables.compoundPersistenceEnabled === "true");
-
-	if (!needsPersistenceArtifacts) {
-		return;
-	}
-
-	await withEphemeralScaffoldNodeModules(targetDir, async () => {
-		if (templateId === "persistence") {
-			await syncPersistenceRestArtifacts({
-				apiTypesFile: path.join("src", "api-types.ts"),
-				outputDir: "src",
-				projectDir: targetDir,
-				variables,
-			});
-			return;
-		}
-
-		await syncPersistenceRestArtifacts({
-			apiTypesFile: path.join("src", "blocks", variables.slugKebabCase, "api-types.ts"),
-			outputDir: path.join("src", "blocks", variables.slugKebabCase),
-			projectDir: targetDir,
-			variables,
-		});
-	});
-}
-
-/**
- * Locate a node_modules directory containing `typia` relative to the project
- * tools package root.
- *
- * Search order:
- * 1. `PROJECT_TOOLS_PACKAGE_ROOT/node_modules`
- * 2. The monorepo root resolved from `PROJECT_TOOLS_PACKAGE_ROOT`
- * 3. The monorepo root `node_modules`
- *
- * @returns The first matching path, or `null` when no candidate contains `typia`.
- */
-function resolveScaffoldGeneratorNodeModulesPath(): string | null {
-	const candidates = [
-		path.join(PROJECT_TOOLS_PACKAGE_ROOT, "node_modules"),
-		path.resolve(PROJECT_TOOLS_PACKAGE_ROOT, "..", ".."),
-		path.resolve(PROJECT_TOOLS_PACKAGE_ROOT, "..", "..", "node_modules"),
-	];
-
-	for (const candidate of candidates) {
-		if (fs.existsSync(path.join(candidate, "typia", "package.json"))) {
-			return candidate;
-		}
-	}
-
-	return null;
-}
-
-/**
- * Temporarily symlink a scaffold generator node_modules directory into the
- * target project while running an async callback.
- *
- * The helper resolves the source path via `resolveScaffoldGeneratorNodeModulesPath()`
- * and uses `EPHEMERAL_NODE_MODULES_LINK_TYPE` for the symlink. The temporary
- * link is removed in the `finally` block so cleanup still happens if the
- * callback throws.
- *
- * @param targetDir Absolute scaffold target directory.
- * @param callback Async work that requires a resolvable `node_modules`.
- * @returns A promise that resolves after the callback and cleanup complete.
- */
-async function withEphemeralScaffoldNodeModules(
-	targetDir: string,
-	callback: () => Promise<void>,
-): Promise<void> {
-	const targetNodeModulesPath = path.join(targetDir, "node_modules");
-	if (fs.existsSync(targetNodeModulesPath)) {
-		await callback();
-		return;
-	}
-
-	const sourceNodeModulesPath = resolveScaffoldGeneratorNodeModulesPath();
-	if (!sourceNodeModulesPath) {
-		throw new Error(
-			"Unable to resolve a node_modules directory with typia for scaffold-time REST artifact generation.",
-		);
-	}
-
-	await fsp.symlink(sourceNodeModulesPath, targetNodeModulesPath, EPHEMERAL_NODE_MODULES_LINK_TYPE);
-
-	try {
-		await callback();
-	} finally {
-		await fsp.rm(targetNodeModulesPath, { force: true, recursive: true });
-	}
-}
-
-async function normalizePackageManagerFiles(
-	targetDir: string,
-	packageManagerId: PackageManagerId,
-): Promise<void> {
-	const yarnRcPath = path.join(targetDir, ".yarnrc.yml");
-
-	if (packageManagerId === "yarn") {
-		await fsp.writeFile(yarnRcPath, "nodeLinker: node-modules\n", "utf8");
-		return;
-	}
-
-	if (fs.existsSync(yarnRcPath)) {
-		await fsp.rm(yarnRcPath, { force: true });
-	}
-}
-
-async function normalizePackageJson(targetDir: string, packageManagerId: PackageManagerId): Promise<void> {
-	const packageJsonPath = path.join(targetDir, "package.json");
-	if (!fs.existsSync(packageJsonPath)) {
-		return;
-	}
-
-	const packageManager = getPackageManager(packageManagerId);
-	const packageJson = JSON.parse(await fsp.readFile(packageJsonPath, "utf8")) as {
-		packageManager?: string;
-		scripts?: Record<string, string>;
-	};
-	packageJson.packageManager = packageManager.packageManagerField;
-
-	if (packageJson.scripts) {
-		for (const [key, value] of Object.entries(packageJson.scripts)) {
-			if (typeof value === "string") {
-				packageJson.scripts[key] = transformPackageManagerText(value, packageManagerId);
-			}
-		}
-	}
-
-	await fsp.writeFile(packageJsonPath, `${JSON.stringify(packageJson, null, "\t")}\n`, "utf8");
-}
-
-async function removeUnexpectedLockfiles(
-	targetDir: string,
-	packageManagerId: PackageManagerId,
-): Promise<void> {
-	const keep = new Set(LOCKFILES[packageManagerId] ?? []);
-	const allLockfiles = Object.values(LOCKFILES).flat();
-
-	await Promise.all(
-		allLockfiles.map(async (filename) => {
-			if (keep.has(filename)) {
-				return;
-			}
-
-			const filePath = path.join(targetDir, filename);
-			if (fs.existsSync(filePath)) {
-				await fsp.rm(filePath, { force: true });
-			}
-		}),
-	);
-}
-
-async function defaultInstallDependencies({
-	projectDir,
-	packageManager,
-}: InstallDependenciesOptions): Promise<void> {
-	execSync(formatInstallCommand(packageManager), {
-		cwd: projectDir,
-		stdio: "inherit",
-	});
-}
-
-function isOfficialWorkspaceProject(projectDir: string): boolean {
-	const packageJsonPath = path.join(projectDir, "package.json");
-	if (!fs.existsSync(packageJsonPath)) {
-		return false;
-	}
-
-	const packageJson = JSON.parse(fs.readFileSync(packageJsonPath, "utf8")) as GeneratedPackageJson;
-	return (
-		packageJson.wpTypia?.projectType === "workspace" &&
-		packageJson.wpTypia?.templatePackage === OFFICIAL_WORKSPACE_TEMPLATE_PACKAGE
-	);
-}
-
-async function applyWorkspaceMigrationCapability(
-	projectDir: string,
-	packageManager: PackageManagerId,
-): Promise<void> {
-	const packageJsonPath = path.join(projectDir, "package.json");
-	const packageJson = JSON.parse(
-		await fsp.readFile(packageJsonPath, "utf8"),
-	) as GeneratedPackageJson;
-	const wpTypiaPackageVersion = getPackageVersions().wpTypiaPackageVersion;
-	const canonicalCliSpecifier =
-		wpTypiaPackageVersion === "^0.0.0"
-			? "wp-typia"
-			: `wp-typia@${wpTypiaPackageVersion.replace(/^[~^]/u, "")}`;
-	const migrationCli = (args: string) =>
-		formatPackageExecCommand(packageManager, canonicalCliSpecifier, `migrate ${args}`);
-
-	packageJson.scripts = {
-		...(packageJson.scripts ?? {}),
-		"migration:init": migrationCli("init --current-migration-version v1"),
-		"migration:snapshot": migrationCli("snapshot"),
-		"migration:diff": migrationCli("diff"),
-		"migration:scaffold": migrationCli("scaffold"),
-		"migration:doctor": migrationCli("doctor --all"),
-		"migration:fixtures": migrationCli("fixtures --all"),
-		"migration:verify": migrationCli("verify --all"),
-		"migration:fuzz": migrationCli("fuzz --all"),
-	};
-
-	await fsp.writeFile(packageJsonPath, `${JSON.stringify(packageJson, null, "\t")}\n`, "utf8");
-
-	writeMigrationConfig(projectDir, {
-		blocks: [],
-		currentMigrationVersion: "v1",
-		snapshotDir: "src/migrations/versions",
-		supportedMigrationVersions: ["v1"],
-	});
-	ensureMigrationDirectories(projectDir, []);
-	writeInitialMigrationScaffold(projectDir, "v1", []);
-}
-
 export async function scaffoldProject({
 	projectDir,
 	templateId,
@@ -866,7 +589,7 @@ export async function scaffoldProject({
 	}
 
 	try {
-		await ensureDirectory(projectDir, allowExistingDir);
+		await ensureScaffoldDirectory(projectDir, allowExistingDir);
 		await copyInterpolatedDirectory(
 			templateSource.templateDir,
 			projectDir,
